feat(v0.8.0): "Clarity" release with modernized syntax
## Language Changes

### New Preferred Keywords
- `gen` replaces `gene` for genome declarations
- `rule` replaces `constraint` for constraint declarations
- `evo` replaces `evolves` for evolution declarations
- `docs` replaces `exegesis` for documentation blocks

### Modernized Type Names (Rust-aligned)
- `string` replaces `String`
- `bool` replaces `Bool`
- `i8/i16/i32/i64` replace `Int8/Int16/Int32/Int64`
- `u8/u16/u32/u64` replace `UInt8/UInt16/UInt32/UInt64`
- `f32/f64` replace `Float32/Float64`
- `()` replaces `Void` for unit type
- `Vec<T>` replaces `List<T>` for vectors

### New Features
- Tree shaking for dead code elimination
- Migration tool: `dol-migrate 0.7-to-0.8`
- CLI flags: --tree-shake, --roots, --shake-analyze

### Internal Changes
- Lexer updated with new keywords and type tokens
- Parser supports both old and new syntax
- Deprecation warnings with source locations

+++ tests/codegen/golden/expected/gene_with_fields.rs
@@ -1,12 +1,12 @@
 #[derive(Debug, Clone, PartialEq)]
 pub struct Container {
-    pub id: isize,
+    pub id: u64,
     pub name: String,
     pub running: bool,
 }
 
 impl Container {
-    pub fn new(id: isize, name: String, running: bool) -> Self {
+    pub fn new(id: u64, name: String, running: bool) -> Self {
         Self {
             id,
             name,
